example_django_app/requirements.txt
Django==1.11.6
django-autofixture==0.12.1
djangorestframework==3.7.0
pytz==2017.2
